move hello-world component files to a directory
