+++ src/components/index.tsx
@@ -1,1 +1,1 @@
-export * from './hello-world';
+export * from './hello-world/hello-world';
